+++ CHANGELOG.md
@@ -38,8 +38,11 @@ Please complete the following steps before upgrading Cumulus.
     reusing their existing security group, ensuring compatibility with existing
     `data-persistence-tf` and `cumulus-tf` modules.
 - **CUMULUS-4473**
-  - Bulk Operations API endpoints now support Granule Inventory Reports and `s3GranuleIdInputFile`.
-  - Updated Bulk Operations API endpoints to accept a list of granuleIds instead of granule objects in the payload.
+  - Updated Granules Bulk Operations API endpoints to:
+    - Support `granuleInventoryReportName` and `s3GranuleIdInputFile` in the payload.
+    - Accept a list of granuleIds instead of granule objects in the payload.
+    - Return consistent output formats across endpoints (previously, some endpoints aggregated errors
+      while others returned per-granule errors)
   - Updated `/executions/search-by-granules` and `/executions/workflows-by-granules` endpoints to accept
     granuleIds instead of granule objects in the payload.